multi: use sqldb v2 in litd
This commit updates litd to use the new sqldb v2 package. Note that this
with just this commit, litd will not utilize the capabilities of sqldb
v2 to run specific post migrations steps (such as migrating the kvdb to
SQL). That functionality will be added in later commits.

Instead, this commit just focuses on adding support for the new
sqldb v2 package, and the functionality of the SQL stores are expected
to remain the same as prior to this commit.