firewalldb: use queries to assert migration results
As the firewalldb package kvdb to sql migration tests creates `sqlc`
models to assert the migration results, we will need to update those
call sites to instead use the `sqlcmig6` models instead, in order to
be compatible with the `SQLMig6Queries` queries.

However, since we can't update the `SQLDB` methods to use `sqlcmig6`
models as params, we need to update the test code assertion to instead
use the `SQLQueries` object directly instead of the `SQLDB` object. This
makes it easy to swap that `SQLQueries` object to a `SQLMig6Queries`
object in the commit that updates the firewalldb package to use the
`sqlcmig6` package for the kvdb to sql migration.