firewalldb: migrate privacy pairs one at a time
Update the privacy mapper migration to iterate over each individual
privacy pair and insert and validate each pair at the time, instead of
collecting all pairs and then inserting them in bulk. This is done to
minimize the memory usage during the migration, as the privacy mapper
store can contain a large number of pairs.